test: Extract FakeTrivialConnection
