+++ master/buildbot/test/unit/test_worker_marathon.py
@@ -22,31 +22,9 @@ from buildbot.process.properties import Properties
 from buildbot.test.fake import fakebuild
 from buildbot.test.fake import fakemaster
 from buildbot.test.fake import httpclientservice as fakehttpclientservice
+from buildbot.test.fake.fakeprotocol import FakeTrivialConnection as FakeBot
 from buildbot.test.util.misc import TestReactorMixin
-from buildbot.util import subscription
 from buildbot.worker.marathon import MarathonLatentWorker
-
-
-class FakeBot:
-    info = {}
-
-    def __init__(self):
-        self._disconnectSubs = subscription.SubscriptionPoint("disconnections from FakeBot")
-
-    def notifyOnDisconnect(self, cb):
-        return self._disconnectSubs.subscribe(cb)
-
-    def waitForNotifyDisconnectedDelivered(self):
-        return self._disconnectSubs.waitForDeliveriesToFinish()
-
-    def notifyDisconnected(self):
-        self._disconnectSubs.deliver()
-
-    def loseConnection(self):
-        self.notifyDisconnected()
-
-    def remoteSetBuilderList(self, builders):
-        return defer.succeed(None)
 
 
 class TestMarathonLatentWorker(unittest.TestCase, TestReactorMixin):
